+++ ipatests/util.py
@@ -39,12 +39,11 @@ import ldap.modlist
 
 import ipalib
 from ipalib import api
+from ipalib.install.kinit import kinit_keytab, kinit_password
 from ipalib.plugable import Plugin
 from ipalib.request import context
 from ipapython.dn import DN
-from ipapython.ipautil import (
-    private_ccache, kinit_password, kinit_keytab, run
-)
+from ipapython.ipautil import private_ccache, run
 from ipaplatform.paths import paths
 
 if six.PY3:
